telegram.js (UX): Add lazy loading support
It's painful to wait for 10000 messages to be loaded at a time, the
application's latency will be high for such an activity.

As such, this commit added lazy loading support. This means we can
split the load into several chunks. The main purpose is that don't
let the user wait for too long to see the first message to be shown!

In this commit we introduce 3 types of load.
  1) `MSG_LOAD_PREPEND`
  2) `MSG_LOAD_APPEND`
  3) `MSG_LOAD_REPLACE`

The load action will depend on what load type is. By default if the
type isn't specified, it will use `MSG_LOAD_PREPEND`.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>